image_checker/ai-image-detector/ai-image-detector-1/requirements.txt
numpy
pillow
tensorflow
scikit-learn
opencv-python
matplotlib
requests